Added step to configure `example.config` to `bin/config`
Tried to ease up the kickstart section. I ran into the problem following the instruction but nothing happend when I clicked on `b3.exe`.
After thinking and searching, I found the missing step that a configuration needs to be available in the same path of the `b3.exe`.
I am hoping that with this added step, the start and the usage of b3 will be increased.